% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_tutorials}
\alias{list_tutorials}
\title{List available tutorials}
\usage{
list_tutorials()
}
\value{
A character vector of tutorial file names (invisibly).
}
\description{
Lists the tutorial R Markdown files included with the package.
Tutorials are designed for classroom use in Korean political science
methods courses. Each tutorial is available in two formats:
\enumerate{
\item Plain Rmd for editing in RStudio (\code{\link{open_tutorial}})
\item Interactive \pkg{learnr} format (\code{\link{run_tutorial}})
}
}
\examples{
list_tutorials()

}
