% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_tutorial}
\alias{open_tutorial}
\title{Open a tutorial file}
\usage{
open_tutorial(name, dest_dir = getwd())
}
\arguments{
\item{name}{Tutorial name (with or without .Rmd extension), or a number
corresponding to the tutorial order (1-9).}

\item{dest_dir}{Directory to copy the file to. Defaults to the current
working directory.}
}
\value{
The path to the copied file (invisibly).
}
\description{
Copies a tutorial R Markdown file to the specified directory (default:
current working directory) so students can edit and run it in RStudio.
}
\examples{
\donttest{
# Copy by name
open_tutorial("01-tidyverse-basics")

# Copy by number
open_tutorial(1)
}

}
\seealso{
\code{\link{run_tutorial}} for the interactive browser version.
}
