% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_tutorial}
\alias{run_tutorial}
\title{Run an interactive tutorial}
\usage{
run_tutorial(name)
}
\arguments{
\item{name}{Tutorial name or number (1-6). Use \code{\link{list_tutorials}}
to see available tutorials.}
}
\description{
Launches a \pkg{learnr} interactive tutorial in the browser. Students
can type and run code directly in the browser with hints and solutions.
Requires the \pkg{learnr} package.
}
\examples{
\dontrun{
run_tutorial(1)  # Launch tutorial 1 in browser
}

}
\seealso{
\code{\link{open_tutorial}} for the plain Rmd version.
}
