% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{assemblykor-package}
\alias{assemblykor}
\alias{assemblykor-package}
\title{assemblykor: Korean National Assembly Data for Political Science Education}
\description{
Provides ready-to-use datasets from the Korean National Assembly for
teaching quantitative methods in political science. Includes seven
built-in datasets covering legislator metadata, bills, asset
declarations, policy seminars, committee speeches, plenary vote
tallies, and member-level roll call votes.
}
\section{Built-in datasets}{

\itemize{
\item \code{\link{legislators}}: 947 MP records (20th-22nd assemblies)
\item \code{\link{bills}}: 60,925 legislative bills
\item \code{\link{wealth}}: 2,928 legislator-year asset declarations
\item \code{\link{seminars}}: 5,962 legislator-year seminar records
\item \code{\link{speeches}}: 15,843 speech records (22nd, Science & ICT Committee)
\item \code{\link{votes}}: 8,050 plenary vote tallies (20th-22nd assemblies)
\item \code{\link{roll_calls}}: 383,739 member-level roll call votes (22nd assembly)
}
}

\section{Download functions}{

\itemize{
\item \code{\link{get_bill_texts}}: 60,925 bill propose-reason texts
\item \code{\link{get_proposers}}: 769,773 co-sponsorship records
\item \code{\link{get_speech_tokens}}: 665,055 morpheme tokens for
the \code{speeches} dataset (Kiwi morphological analyzer)
}
}

\section{Tutorials}{

Nine Korean-language tutorials covering tidyverse, visualization, regression,
panel data, text analysis, network analysis, roll call analysis, bill success,
and speech patterns. Use \code{\link{list_tutorials}} to see all tutorials,
and \code{\link{open_tutorial}} to copy them to your working directory.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://kyusik-yang.github.io/assemblykor/}
  \item \url{https://github.com/kyusik-yang/assemblykor}
  \item Report bugs at \url{https://github.com/kyusik-yang/assemblykor/issues}
}

}
\author{
\strong{Maintainer}: Kyusik Yang \email{kyusik.yang@nyu.edu}

}
